chore:split redux types and update logic links
