+++ client/src/redux/gameState/gameStateReducer.ts
@@ -1,5 +1,5 @@
-import { CLEAR_MISINFO, DEBUNK_MISINFO, DISCARD_ACTION, GameStateActionTypes, LOG_ON_OFF, MOVE_ACTION, SHARE_CARD, UPDATE_GAME_STATE } from './../../types/gameStateTypes';
-import { gameState } from '../../socket-io-client/dummy-state'
+import { CLEAR_MISINFO, DEBUNK_MISINFO, DISCARD_ACTION, GameStateActionTypes, LOG_ON_OFF, MOVE_ACTION, SHARE_CARD, UPDATE_GAME_STATE } from './reduxTypes';
+import { initDummyState as gameState } from '../../logic/dummyState.REDO_CO'
 import { Gamestate } from '../../types/gameStateTypes'
 import { clearMisinfo, debunkMisinfo, discardCard, logOnOff, moveAction, shareCard } from '../../logic/actions.newState_CO'
 // import {emit} from '../backend-dummy-client/dummy-client'
